Refactored code to remove circular imports by placing the ten_percent_function in the augment.py file and by refactoring the augment functions to take in a boolean crop parameter
